Refactored to isolate C and C++ specializations
